DOM EVENTS EXCERSISE 1
1.If you click on the list item, it toggles the .done class on and off.

i did this by creating a for loop to add event listeners to each li
item and call my liClick function. the liClick function then toggles
the done class.

It took me longer than i would like to admit to figure this out and alot
of research outside of the course curriculum to be able to figure this
out i also had to research the 'this' keyword which is taught further
along in the course however i think i understand it now and it will
help me in the future as it has helped me to exercise my reasearching
skills.