FINISHED DOM EVENTS TASk
1) If you click on the list item, it toggles the .done class on and off.

2) Add buttons next to each list item to delete the item when clicked on
its corresponding delete button.

3) BONUS: When adding a new list item, it automatically adds the delete
button next to it (hint: be sure to check if new items are clickable
too!)

This took me way to long to figure out and not sure if i would be able
to do it again. i had to search through overs code on stack overflow and
and modify things to complete the task. took me a few days to get my
head around it but has definatly improved my reasearch skills. my
original code to complete the first task had to be completely removed
had to redo the code with the intention of completeing the next tasks
rather than just adding to it.